+++ src/main/java/roomescape/controller/admin/AdminReservationController.java
@@ -10,9 +10,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.Reservation;
-import roomescape.exception.NotFoundException;
-import roomescape.request.ReservationRequest;
-import roomescape.response.ReservationResponse;
+import roomescape.dto.request.ReservationRequest;
+import roomescape.dto.response.ReservationResponse;
 import roomescape.service.ReservationService;
 
 import java.net.URI;
@@ -40,8 +39,8 @@ public class AdminReservationController {
             ReservationResponse reservationResponse = ReservationResponse.from(reservationReturned);
 
             return ResponseEntity.created(getLocation(reservationResponse.id())).body(reservationResponse);
-        } catch (NotFoundException e) {
-            return ResponseEntity.notFound().build();
+        } catch (IllegalArgumentException e) {
+            return ResponseEntity.badRequest().build();
         }
     }
