+++ src/main/java/roomescape/controller/view/AdminViewController.java
@@ -3,9 +3,9 @@ package roomescape.controller.view;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import roomescape.response.ReservationResponse;
-import roomescape.response.ReservationTimeResponse;
-import roomescape.response.ThemeResponse;
+import roomescape.dto.response.ReservationResponse;
+import roomescape.dto.response.ReservationTimeResponse;
+import roomescape.dto.response.ThemeResponse;
 import roomescape.service.ReservationService;
 import roomescape.service.ReservationTimeService;
 import roomescape.service.ThemeService;
