+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -1,10 +1,9 @@
 package roomescape.service;
 
+import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.domain.ReservationTime;
 import roomescape.repository.ReservationTimeRepository;
-
-import java.util.List;
 
 @Service
 public class ReservationTimeService {
